SiaPrime v1.4.0 RC1
Release candidate one for SiaPrime v1.4.0. All tests passing, syncs
with the blockchain, and can send SCP. Currently not abot to create
new file contracts for hosting but I suspect it is because there
are no SiaPrime v1.4.0 hosts on the networkor because I was trying
to use the SiaPrime v1.3.8 UI to do so as the hosting and renting
tests pass.