Preliminary (draft) implementation for SPF hardfork
The claim generated by given SiafundOutput is calculated in this way:
(CurrentSiafundPool - SiafundPoolAtLastClaimForThisOutput) / TotalSiafundCount * OutputAmount.
We calculate total siafund income since the output 'started' and divide by total siafund
count. Then we multiply the result by amount of siafunds in this output.
On SPF Emission Hardfork the amount of siafunds is changed, so we slightly modify the formula.
If ClaimStart is after the hardfork, we just use the same formula but with new amount of siafunds.
If ClaimStart was before the hardfork, we divide the difference between HardforkPool and ClaimStart by
old amount of siafunds, and the difference between CurrentPool and HardforkPool is divided by new amount.

Signed-off-by: Pavel Dolgov <pdolgov99@gmail.com>