Web-Based Wallet GUI RC3
Adds a web-based GUI module that is loaded in with the `u` modules
parameter. When enabled it is accessed on localhost at the same port
that the wallet API is set up for; by default this is
`http://localhost:4280`. All sensitive interaction with the wallet
GUI requires the user to also supply their session ID (which they are
assigned after supplying a valid wallet password) to protect the
user from accidentally compromising their seed, password, or coins.

All static resource files are stored in the `./node/api/resources/`
directory which should make it easy to adjust graphics and css
styling in the future.

Adds a downloader module that is loaded with the `d` modules parameter.
When enabled and the wallet has no consensus the downloader module will
download the bootstrap consensus from scpri.me before loading the
consensus module.

Adds a scp-ui package that behaves as a native application. The Makefile,
build scripts, and release scripts have all been updated to compile the
scp-ui application.