Add CLI To Send ScPrime Coins Mapped By A CSV RC2
The following command will send SCP to multiple addresses as
mapped by the supplied input CSV file.

```
./spc wallet send scp_from_csv "/path/to/file.csv"
```

Long Description:

Send scprimecoins to multiple addresses in the same transaction
as defined in the supplied CSV file. The first column must be the
amount of scprimecoin to send specified in units, e.g. 1.23KS.
Run 'wallet --help' for a list of units. If no unit is supplied,
hastings will be assumed. The second column must be the 76-byte
hexadecimal address.

A dynamic transaction fee is applied depending on the size of the
transaction and how busy the network is.